requirements.txt
google-genai>=0.2.0
python-dotenv>=1.0.1
scikit-learn>=1.0.0
groq>=0.4.0
